Feat : DM 새 메시지 알림 추가 및 홍보게시판 삭제 수정
MessageService
- notifyUser 메서드 추가
  웹소캣을 통해 수신자에게 새 메시지 알림 메시지를 보냄

PromotionPostService
- 게시글 삭제 시 첨부파일이 있는지 확인한 뒤
  S3 이미지 삭제 로직 실행되게 수정

+++ src/main/java/aespa/groovymap/promotionpost/service/PromotionPostService.java
@@ -279,20 +279,22 @@ public class PromotionPostService {
         savedPosts.forEach(savedPostRepository::delete);
 
         // S3에 업로드된 파일 삭제 로직 추가
-        List<String> fileNames = promotionPost.getImageSet().stream()
-                .map(image -> image.getFileName())
-                .collect(Collectors.toList());
-        String filesToDelete = String.join(",", fileNames);
-        Map<String, Boolean> deleteResults = upDownService.removeFile(filesToDelete);
+        if (!promotionPost.getImageSet().isEmpty()) {
+            List<String> fileNames = promotionPost.getImageSet().stream()
+                    .map(image -> image.getFileName())
+                    .collect(Collectors.toList());
+            String filesToDelete = String.join(",", fileNames);
+            Map<String, Boolean> deleteResults = upDownService.removeFile(filesToDelete);
 
-        // 삭제 결과 로그 출력
-        deleteResults.forEach((fileName, success) -> {
-            if (success) {
-                log.info("Deleted file from S3: {}", fileName);
-            } else {
-                log.error("Failed to delete file from S3: {}", fileName);
-            }
-        });
+            // 삭제 결과 로그 출력
+            deleteResults.forEach((fileName, success) -> {
+                if (success) {
+                    log.info("Deleted file from S3: {}", fileName);
+                } else {
+                    log.error("Failed to delete file from S3: {}", fileName);
+                }
+            });
+        }
 
         // 게시글 삭제
         promotionPostRepository.delete(promotionPost);
